moving cookie creation and auth login to helper functions
