+++ lib/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable */
 const fs = require('fs');
+const { v4: uuid } = require('uuid');
 const request = require('request-promise-native');
 const {
   buildDbFunctions,
@@ -9,8 +10,19 @@ const {
   START_QUERY_PARAMETER_KEY,
   SEARCH_ALL_PAGES,
   UI_CONSTANTS,
-  OAUTH
+  OAUTH,
+  MIDDLEWARE,
+  COOKIES
 } = CONSTANTS;
+
+const {
+  REDIRECT_URL
+} = MIDDLEWARE;
+
+const {
+  KEY: COOKIE_KEY,
+  COOKIE_MAX_AGE
+} = COOKIES;
 
 const readFile = path => new Promise((resolve, reject) => {
   fs.readFile(path, (error, buffer) => {
@@ -486,6 +498,26 @@ const getGmailProfile = async (token) => {
   }
 };
 
+const getAuthURLAndSendRedirectJSON = (res, contextObject, statusCode = 200) => {
+  const { credentialsObject, id, email } = contextObject;
+  const authURL = getAuthUrlFromCredentials(credentialsObject, id, email);
+  return res.status(statusCode).json({ [REDIRECT_URL]: authURL });
+};
+
+const createSessionCookie = async (res, redisFunctions) => {
+  const cookieValue = uuid();
+  res.cookie(COOKIE_KEY, cookieValue, { maxAge: COOKIE_MAX_AGE, httpOnly: true });
+
+  try {
+    await redisFunctions.mapCookieAndUserId(cookieValue, userId);
+  } catch (err) {
+    console.error('error setting key/value in redis:', err);
+    return Promise.reject(err);
+  }
+
+  return cookieValue;
+};
+
 const exportsObject = {
   readFile,
   readJson,
@@ -502,7 +534,9 @@ const exportsObject = {
   requestToken,
   refreshToken,
   getGmailProfile,
-  getAuthUrlFromCredentials
+  getAuthUrlFromCredentials,
+  getAuthURLAndSendRedirectJSON,
+  createSessionCookie
 };
 
 module.exports = exportsObject;
